+++ avaliacao.hpp
@@ -12,18 +12,18 @@ private:
     int _id;                   // id da avaliacao
     int _alvoId;              // id da disciplina/professor/turma
     std::string _tipo;        // disciplina, professor ou turma
-    int _nota;                // 0 a 10
+    double _nota;                // 0 a 5
     std::string _comentario;  // comentario do avaliador
     std::string _data;        // data da avaliacao
 
 public:
     Avaliacao();
-    Avaliacao(int id, int alvoId, const std::string &tipo, int nota, const std::string &comentario, const std::string &data);
+    Avaliacao(int id, int alvoId, const std::string &tipo, double nota, const std::string &comentario, const std::string &data);
 
     int getId() const;
     int getAlvoId() const;
     std::string getTipo() const;
-    int getNota() const;
+    double getNota() const;
     std::string getComentario() const;
     std::string getData() const;
 };
